# FTI - Optimize `SceneTree` traversal