@lawnjelly 2025-05-24
> We should probably remove the Legacy codepath before 4.5 is released unless there's a good reason to keep it selectable as an option. The change should be 100% transparent for users anyway (nonwithstanding bugs).

Yup we should be able to, it's just there as a temporary hedge for regressions really.