@TranquilMarmot 2025-05-27
@lawnjelly do you know if this would affect how physics interpolation would interact with a `VehicleBody3D`?

Building the latest `master`, I have a `VehicleBody3D` that does not move properly after this PR was merged.  Turning off physics interpolation it works, and checking out the commit before this was merged it works.

It seems like the `RigidBody3D` that represents the `VehicleBody3D` is updating its position properly, but the visual components of it are not updating.